Extracts UIImage test helper to a separate file
